requirements.txt
# AP2 Financial Reports Scraper - Requirements
# Python 3.8+ required

# Web Scraping
selenium>=4.15.0
undetected-chromedriver>=3.5.4
beautifulsoup4>=4.12.0
lxml>=4.9.0

# PDF Processing - Enhanced table extraction
PyMuPDF>=1.23.0  # fitz - Primary table extraction
pdfplumber>=0.10.0  # Text extraction fallback
camelot-py[cv]>=0.11.0  # Advanced table detection
PyPDF2>=3.0.0  # PDF utilities
tabula-py>=2.9.0  # Alternative table extraction

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.0

# Optional but recommended
# For better PDF parsing on complex documents
opencv-python>=4.8.0
pytesseract>=0.3.10
